requirements.txt
Django==5.1
djangorestframework==3.14.0
django-cors-headers==4.2.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Pillow==10.1.0
